Generate source maps for es5 and es6
