+++ lib/browser.js
@@ -2,7 +2,7 @@
 
 require('setimmediate')
 
-self.Storj = require('./index.js');
+self.Storj = require('./storj.js');
 self.Storj.DEFAULTS.store = new (require('memory-blob-store'))();
 
 module.exports = self.Storj;
